Finalized decoupling logic, uopdated class names
