+++ Sources/RoboKit/Socket/Client/States/ClientStateMethods.swift
