+++ Sources/RoboKit/Socket/Connection/Messages/ConnectionMessageMethods.swift
