+++ Sources/RoboKit/Socket/Connection/States/ConnectionStateMethods.swift
