+++ Sources/RoboKit/Socket/JSON/Manager/CodingManager.swift
