+++ Sources/RoboKit/Socket/JSON/Models/JSONErrors.swift
