+++ Sources/RoboKit/Socket/JSON/Models/MessageModel.swift
